Initial commit – Transport_Company_App_UI

+++ src/main/java/com/transportcompany/dto/TransportDTO.java
@@ -1,8 +1,6 @@
 package com.transportcompany.dto;
 
 import lombok.*;
-
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Getter
@@ -11,8 +9,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @Builder
 public class TransportDTO {
-
-    private int id;
+    private Integer id;
     private String startPoint;
     private String endPoint;
     private LocalDate departureDate;
@@ -22,9 +19,9 @@ public class TransportDTO {
     private Double cargoWeight;
     private boolean paid;
 
-    private int clientId;
-    private int employeeId;
-    private int vehicleId;
-    private int companyId;
-    private BigDecimal priceId;
+    private Integer clientId;
+    private Integer employeeId;
+    private Integer vehicleId;
+    private Integer companyId;
+    private Integer priceId;
 }
